implemented Navigation Drawer
