+++ app/src/main/AndroidManifest.xml
@@ -45,6 +45,11 @@
             android:name="com.google.android.gms.ads.AdActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|uiMode|screenSize|smallestScreenSize"
             android:theme="@android:style/Theme.Translucent" />
+        
+        <activity android:name=".MenuActivity"
+            android:screenOrientation="portrait">
+
+        </activity>
 
         <service android:name=".NotiService"
             android:enabled="true" />
