+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <group android:checkableBehavior="single">
+        <item
+            android:id="@+id/nav_weight"
+            android:title="@string/navigation_drawer_weight_graph" />
+        <item
+            android:id="@+id/nav_damage"
+            android:title="Schaden melden" />
+    </group>
+
+    <item android:title="Profil">
+        <menu>
+            <item
+                android:id="@+id/nav_update"
+                android:title="Profil bearbeiten" />
+            <item
+                android:id="@+id/nav_logout"
+                android:title="Abmelden" />
+        </menu>
+    </item>
+
+</menu>
