+++ app/src/main/res/values-de/strings.xml
@@ -86,4 +86,12 @@
     <string name="menu_sync">Synchronisieren</string>
     <string name="menu_download">Lade Einträge</string>
     <string name="menu_logout">Abmelden</string>
+
+    <string name="navigation_drawer_open">Navigation Drawer öffnen</string>
+    <string name="navigation_drawer_close">Navigation Drawer schließen</string>
+
+    <string name="navigation_drawer_weight_graph">Gewichts Diagramm</string>
+
+
+    <string name="action_settings">Settings</string>
 </resources>
